# src/swagger.yaml
swagger: "2.0"
info:
  description: This is Doran server.
  version: "1.0.0"
  title: Swagger Doran
basePath: /api/v1
tags:
- name: doran
  description: Everything about Doran
schemes:
- http
paths:
  /register:
    post:
      tags:
      - doran
      summary: register
      operationId: register
      parameters:
      - in: body
        name: body
        description: User register object
        schema:
          type: object
          properties:
            phoneId:
              type: string
            nickname:
              type: string
            age:
              type: integer
            sex:
              type: string
            location:
              type: string
        required: true
      responses:
        400:
          description: Bad Request
        500:
          description: Internal Server Error
  /login:
    post:
      tags:
      - doran
      summary: login
      operationId: login
      parameters:
      - in: body
        name: body
        description: User login object
        schema:
          type: object
          properties:
            phoneId:
              type: string
        required: true
      responses:
        400:
          description: Bad Request
        401:
          description: Unauthorized
  /boards:
    get:
      tags:
      - doran
      summary: read boards
      operationId: readBoards
      parameters:
      - in: query
        name: skip
        description: number of pages to skip
        type: integer
        required: true
      - in: query
        name: limit
        description: number of messages in a page
        type: integer
        required: true  
      responses:
        400:
          description: Bad Request
        401:
          description: Unauthorized
    post:
      tags:
      - doran
      summary: write board
      operationId: writeBoard
      parameters:
      - in: body
        name: body
        description: Board object
        schema:
          type: object
          properties:
            sender:
              type: string
            content:
              type: string
        required: true 
      responses:
        400:
          description: Bad Request  
        401:
          description: Unauthorized
        500:
          description: Internal Server Error
  /rooms:
    get:
      tags:
      - doran
      summary: get rooms
      operationId: getRooms
      parameters:
      - in: query
        name: skip
        description: number of pages to skip
        type: integer
        required: true
      - in: query
        name: limit
        description: number of messages in a page
        type: integer
        required: true
      responses:
        200:
          description: A list of rooms
          schema:
            $ref: '#/components/schemas/Room'
        400:
          description: Bad Request
        401:
          description: Unauthorized
    post:
      tags:
      - doran
      summary: create a room
      operationId: createRoom
      parameters:
      - in: body
        name: body
        description: guest id
        schema:
          type: object
          properties:
            guest:
              type: string
        required: true 
      responses:
        200:
          description: Existed Room
          schema:
            $ref: '#/components/schemas/Room'
        201:
          description: Created Room
          schema:
            $ref: '#/components/schemas/Room'
        400:
          description: Bad Request  
        401:
          description: Unauthorized
        500:
          description: Internal Server Error
  /messages:
    get:
      tags:
      - doran
      summary: get messages
      operationId: getMessages
      parameters:
      - in: query
        name: roomId
        description: room id
        type: string
        required: true
      - in: query
        name: skip
        description: number of pages to skip
        type: integer
        required: true
      - in: query
        name: limit
        description: number of messages in a page
        type: integer
        required: true
      responses:
        400:
          description: Bad Request
        401:
          description: Unauthorized
    post:
      tags:
      - doran
      summary: send a message
      operationId: sendMessage
      parameters:
      - in: body
        name: body
        description: Message object
        schema:
          type: object
          properties:
            roomId:
              type: string
            content:
              type: string
        required: true
      responses:
        400:
          description: Bad Request  
        401:
          description: Unauthorized
        500:
          description: Internal Server Error
components:
  schemas:
    Room:
      type: object
      properties:
        _id:
          type: string
        host:
          type: string
        guest:
          type: string
        createAt:
          type: string
        updatedAt:
          type: string